LibWeb: Use `double` in CSS Token
It was already a double internally, so let's expose that and get rid of
some of clangd's red wiggly underlines.

+++ Userland/Libraries/LibWeb/CSS/Parser/Token.h
@@ -111,7 +111,7 @@ public:
         VERIFY(m_type == Type::Number || m_type == Type::Dimension || m_type == Type::Percentage);
         return m_number_value;
     }
-    float number_value() const
+    double number_value() const
     {
         VERIFY(m_type == Type::Number);
         return m_number_value.value();
@@ -127,14 +127,14 @@ public:
         VERIFY(m_type == Type::Dimension);
         return m_value.bytes_as_string_view();
     }
-    float dimension_value() const
+    double dimension_value() const
     {
         VERIFY(m_type == Type::Dimension);
         return m_number_value.value();
     }
     i64 dimension_value_int() const { return m_number_value.integer_value(); }
 
-    float percentage() const
+    double percentage() const
     {
         VERIFY(m_type == Type::Percentage);
         return m_number_value.value();
@@ -159,7 +159,7 @@ public:
         return token;
     }
 
-    static Token create_number(float value)
+    static Token create_number(double value)
     {
         Token token;
         token.m_type = Type::Number;
@@ -167,7 +167,7 @@ public:
         return token;
     }
 
-    static Token create_percentage(float value)
+    static Token create_percentage(double value)
     {
         Token token;
         token.m_type = Type::Percentage;
